refactor(frontend): split chat stream utilities and session flow (#824)
